requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
httpx==0.27.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.0
tenacity==9.0.0
google-generativeai==0.8.0
aiofiles==24.1.0
prometheus-client==0.21.0
structlog==24.4.0
redis==5.1.0
aiocache==0.12.3